Added GROUP collection to setup.
The STORE collection is used to process a single date: it stores both the data and the geometry. It will be used to extract the geometry and insert it into the DroughtObservatoryMap, ignoring any duplicate key error. Then the data part of the record will be appended to the GROUP collection, so that when the year has been processed, we can group the data and dump the year.